delegate uptime 3-day EMA

+++ delegatelist.go
@@ -2,7 +2,7 @@ package main
 
 type KnownDelegate struct {
 	Id           uint64
-	BlocksStaked uint64
-	BlocksMissed uint64
+	BlocksStaked float64
+	BlocksMissed float64
 	LastHeight   uint64
 }
